% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectLSI.R
\name{calcLSI}
\alias{calcLSI}
\title{Calculate Latent Semantic Indexing (LSI)}
\usage{
calcLSI(mat, ndims = 50, seed.use = 42)
}
\arguments{
\item{mat}{the normalized sparse count matrix}

\item{ndims}{how many dimensions will be calculated, default 50}

\item{seed.use}{the random seed used, default 42}
}
\value{
a list containing the result of svd and parameters used
}
\description{
Perform svd and then LSI to the input count matrix.
}
