% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectLSI.R
\name{projectLSI}
\alias{projectLSI}
\title{Project to Given Latent Semantic Indexing (LSI)}
\usage{
projectLSI(mat, lsi)
}
\arguments{
\item{mat}{the normalized sparce count matrix}

\item{lsi}{a list of the output of calcLSI()}
}
\value{
a dense matrix of svd cell loading
}
\description{
Project the input count matrix to a given LSI space using pre-calculated svd output.
}
