% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectLSI.R
\name{psudoSC}
\alias{psudoSC}
\title{Calculate Psudo-single-cell Count Data}
\usage{
psudoSC(bulk.data, n = 100, depth = 3000)
}
\arguments{
\item{bulk.data}{the bulk RNA-seq counts matrix, genes in row and samples in column.}

\item{n}{how many times to down-sample per bulk sample, default 100}

\item{depth}{how many counts down-sample to, default 3000}
}
\value{
a sparse matrix of down-sampled count matrix with genes in row and n*sample psudo-cells in column
}
\description{
Calculate the psudo-single-cell RNA-seq data by down-sampling bulk RNA-seq counts.
}
